+++ kpops/components/base_components/helm_app.py
@@ -26,12 +26,9 @@ from kpops.components.base_components.kubernetes_app import (
     KubernetesApp,
     KubernetesAppValues,
 )
-from kpops.components.common.kubernetes_model import (
-    K8S_LABEL_MAX_LEN,
-    KubernetesManifest,
-)
 from kpops.config import get_config
-from kpops.manifests.argo import ArgoSyncWave, enrich_annotations
+from kpops.manifest.argo import ArgoSyncWave, enrich_annotations
+from kpops.manifest.kubernetes import K8S_LABEL_MAX_LEN, KubernetesManifest
 from kpops.utils.colorify import magentaify
 from kpops.utils.docstring import describe_attr
 from kpops.utils.pydantic import exclude_by_name
