+++ tests/component_handlers/kubernetes/model.py
@@ -2,7 +2,7 @@ from textwrap import dedent
 
 import pytest
 
-from kpops.components.common.kubernetes_model import KubernetesManifest
+from kpops.manifest.kubernetes import KubernetesManifest
 
 
 class TestKubernetesManifest:
