+++ tests/components/test_helm_app.py
@@ -10,7 +10,7 @@ from kpops.component_handlers.helm_wrapper.model import (
     RepoAuthFlags,
 )
 from kpops.components.base_components.helm_app import HelmApp, HelmAppValues
-from kpops.components.common.kubernetes_model import K8S_LABEL_MAX_LEN
+from kpops.manifest.kubernetes import K8S_LABEL_MAX_LEN
 from kpops.utils.colorify import magentaify
