+++ tests/pipeline/snapshots/test_manifest/test_python_api/manifest.yaml
@@ -1,0 +1,107 @@
+---
+apiVersion: batch/v1
+kind: Job
+metadata:
+  labels:
+    app: resources-manifest-pipeline-my-producer-app
+    chart: producer-app-3.0.3
+    release: resources-manifest-pipeline-my-producer-app
+  name: resources-manifest-pipeline-my-producer-app
+spec:
+  backoffLimit: 6
+  template:
+    metadata:
+      labels:
+        app: resources-manifest-pipeline-my-producer-app
+        release: resources-manifest-pipeline-my-producer-app
+    spec:
+      containers:
+      - env:
+        - name: ENV_PREFIX
+          value: APP_
+        - name: APP_BOOTSTRAP_SERVERS
+          value: http://k8kafka-cp-kafka-headless.kpops.svc.cluster.local:9092
+        - name: APP_SCHEMA_REGISTRY_URL
+          value: http://localhost:8081/
+        - name: APP_OUTPUT_TOPIC
+          value: my-producer-app-output-topic
+        - name: APP_LABELED_OUTPUT_TOPICS
+          value: my-producer-app-output-topic-label=my-labeled-producer-app-topic-output,
+        - name: JAVA_TOOL_OPTIONS
+          value: '-XX:MaxRAMPercentage=75.0 '
+        image: my-registry/my-producer-image:1.0.0
+        imagePullPolicy: Always
+        name: resources-manifest-pipeline-my-producer-app
+        resources:
+          limits:
+            cpu: 500m
+            memory: 2G
+          requests:
+            cpu: 200m
+            memory: 300Mi
+      restartPolicy: OnFailure
+
+---
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  annotations:
+    consumerGroup: my-streams-app-id
+  labels:
+    app: resources-manifest-pipeline-my-streams-app
+    chart: streams-app-3.0.3
+    release: resources-manifest-pipeline-my-streams-app
+  name: resources-manifest-pipeline-my-streams-app
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: resources-manifest-pipeline-my-streams-app
+      release: resources-manifest-pipeline-my-streams-app
+  template:
+    metadata:
+      labels:
+        app: resources-manifest-pipeline-my-streams-app
+        release: resources-manifest-pipeline-my-streams-app
+    spec:
+      containers:
+      - env:
+        - name: ENV_PREFIX
+          value: APP_
+        - name: APP_VOLATILE_GROUP_INSTANCE_ID
+          value: 'true'
+        - name: APP_BOOTSTRAP_SERVERS
+          value: http://k8kafka-cp-kafka-headless.kpops.svc.cluster.local:9092
+        - name: APP_SCHEMA_REGISTRY_URL
+          value: http://localhost:8081/
+        - name: APP_INPUT_TOPICS
+          value: my-input-topic
+        - name: APP_INPUT_PATTERN
+          value: my-input-pattern
+        - name: APP_OUTPUT_TOPIC
+          value: my-output-topic
+        - name: APP_ERROR_TOPIC
+          value: resources-manifest-pipeline-my-streams-app-error
+        - name: APP_LABELED_OUTPUT_TOPICS
+          value: my-output-topic-label=my-labeled-topic-output,
+        - name: APP_LABELED_INPUT_TOPICS
+          value: my-input-topic-label=my-labeled-input-topic,
+        - name: APP_LABELED_INPUT_PATTERNS
+          value: my-input-topic-labeled-pattern=my-labeled-input-pattern,
+        - name: APP_APPLICATION_ID
+          value: my-streams-app-id
+        - name: JAVA_TOOL_OPTIONS
+          value: '-Dcom.sun.management.jmxremote.port=5555 -Dcom.sun.management.jmxremote.authenticate=false
+            -Dcom.sun.management.jmxremote.ssl=false -XX:MaxRAMPercentage=75.0 '
+        image: my-registry/my-streams-app-image:1.0.0
+        imagePullPolicy: Always
+        name: resources-manifest-pipeline-my-streams-app
+        resources:
+          limits:
+            cpu: 500m
+            memory: 2G
+          requests:
+            cpu: 200m
+            memory: 300Mi
+      terminationGracePeriodSeconds: 300
+
