+++ tests/pipeline/test_manifest.py
@@ -3,6 +3,7 @@ from pathlib import Path
 from unittest.mock import ANY, MagicMock
 
 import pytest
+from _pytest.capture import CaptureFixture
 from pytest_mock import MockerFixture
 from pytest_snapshot.plugin import Snapshot
 from typer.testing import CliRunner
@@ -11,8 +12,9 @@ import kpops.api as kpops
 from kpops.cli.main import app
 from kpops.component_handlers.helm_wrapper.helm import Helm
 from kpops.component_handlers.helm_wrapper.model import HelmConfig, Version
-from kpops.components.common.kubernetes_model import KubernetesManifest
 from kpops.const.file_type import PIPELINE_YAML
+from kpops.manifest.kubernetes import KubernetesManifest
+from kpops.utils.yaml import print_yaml
 
 MANIFEST_YAML = "manifest.yaml"
 
@@ -123,7 +125,7 @@ class TestManifest:
         assert result.exit_code == 0, result.stdout
         snapshot.assert_match(result.stdout, MANIFEST_YAML)
 
-    def test_python_api(self):
+    def test_python_api(self, capsys: CaptureFixture, snapshot: Snapshot):
         generator = kpops.manifest_deploy(
             RESOURCE_PATH / "manifest-pipeline" / PIPELINE_YAML,
             environment="development",
@@ -131,11 +133,13 @@ class TestManifest:
         assert isinstance(generator, Iterator)
         resources = list(generator)
         assert len(resources) == 2
-        assert all(
-            isinstance(manifest, KubernetesManifest)
-            for resource in resources
-            for manifest in resource
-        )
+        for resource in resources:
+            for manifest in resource:
+                assert isinstance(manifest, KubernetesManifest)
+                print_yaml(manifest.model_dump())
+
+        captured = capsys.readouterr()
+        snapshot.assert_match(captured.out, MANIFEST_YAML)
 
     def test_streams_bootstrap(self, snapshot: Snapshot):
         result = runner.invoke(
